# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 充值分销开放平台 API
  version: 1.3.0
  description: |
    面向代理商销售渠道服务端的开单、查付、履约、退款申请、账单与结算接口。
    支持平台收款分销，或授权代理自收款后以采购余额支付供货价；后者的 paid 仅代表供货款已扣。
    正式商品仅为 Plus、Pro 5x、Pro 20x、Pro 50x，由 CDK 库存支撑；代理商下单时可选择直接交付 CDK，或在自有品牌页面调用自动直充接口。
    固定 1 元 payment_test 是独立支付联调用途，不属于正式套餐，不签发 CDK、不创建充值任务、不计算佣金、不调用上游。
    自动直充失败时只返回可重新提交标志，不返回明文 CDK；客户页面、品牌和订单入口由代理商自行实现。
    平台收款当前仅支持支付宝，不支持 USDT；代理自收款由代理自行接入自己的支付渠道，不向 Quefa 提交支付宝应用或支付密钥。
    平台不发送邮件通知；自动生成核算结果不等于自动打款，代理出款须人工确认。
    沙箱与生产使用不同域名、数据及密钥。所有接口均执行租户隔离。
    代理商账号默认开放 API，无需申请或预存。采购余额为零时仍可查单及平台代收下单，但不得使用代理自收款/余额采购；管理员明确停用时返回 403。
    余额采购在订单事务内扣供货价，不得透支；已付订单提交履约不重复扣款。
    POST 请求必须携带 Idempotency-Key；签名使用 HTTP 原始请求体字节。
    进度变化通过 fulfillment.updated 通知并可查任务恢复；明确失败或平台确认取消后，只有最新任务 retry_allowed=true 才能重提原订单。
    平台选择退款后重提关闭；代理无主动取消权限，不调用管理员取消或退款执行接口。新增进度通知须部署同版并订阅后生效。
  contact:
    name: 平台技术支持
servers:
  - url: https://tibo.ink
    description: 生产 Partner API（请使用此域名，勿将 POST 请求发往会 301 跳转的旧子域）
...
security:
  - PartnerId: []
    KeyId: []
    Timestamp: []
    Nonce: []
    Signature: []
...
paths:
...
  /v1/redemptions:
    post:
      tags: [CustomerRecharge]
      operationId: createPartnerRedemption
      summary: 代理自建页面提交自动直充或显式 CDK 兑换
      description: 须开通 customRedemptionEnabled，只能使用本代理资源。auto_recharge 使用订单已绑定的 CDK，明文不返回代理前端。202 表示受理，不是成功。同键同原始请求体返回原受理结果；原任务查询用于获取最新状态。收到最近尝试 retry_allowed=true 且 next_action=resubmit 后，允许客户在原订单修正资料并用新幂等键创建下一尝试，无需重新下单或付款；未知结果禁止重提。阶段变化投递 fulfillment.updated，终态沿用 fulfillment.succeeded/failed/cancelled。
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, pattern: '^[a-zA-Z0-9:_.-]{8,120}$'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  additionalProperties: false
                  required: [mode, order_id, credential, customer_confirmed_email]
                  properties:
                    mode: {type: string, const: direct}
                    order_id: {type: string, minLength: 1, maxLength: 100}
                    credential: {$ref: '#/components/schemas/RechargeCredential'}
                    customer_confirmed_email: {type: boolean, const: true}
                - type: object
                  additionalProperties: false
                  required: [mode, order_id, credential, customer_confirmed_email]
                  properties:
                    mode: {type: string, const: auto_recharge, description: 使用订单绑定的 CDK 自动兑换；不向调用方返回明文 CDK}
                    order_id: {type: string, minLength: 1, maxLength: 100}
                    credential: {$ref: '#/components/schemas/RechargeCredential'}
                    customer_confirmed_email: {type: boolean, const: true}
                - type: object
                  additionalProperties: false
                  required: [mode, code, credential, customer_confirmed_email]
                  properties:
                    mode: {type: string, const: cdk}
                    code: {type: string, pattern: '^[A-Z0-9][A-Z0-9_-]{7,63}$', writeOnly: true}
                    credential: {$ref: '#/components/schemas/RechargeCredential'}
                    customer_confirmed_email: {type: boolean, const: true}
      responses:
        '202':
          description: 已受理；queued 不代表充值完成
          content:
            application/json: {schema: {$ref: '#/components/schemas/RedemptionEnvelope'}}
        default: {$ref: '#/components/responses/Error'}
  /v1/redemptions/{redemption_id}:
    get:
      tags: [CustomerRecharge]
      operationId: getPartnerRedemption
      summary: 查询本代理充值任务
      parameters:
        - name: redemption_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: 统一状态，无底层信息
          content:
            application/json: {schema: {$ref: '#/components/schemas/RedemptionEnvelope'}}
        default: {$ref: '#/components/responses/Error'}
...
  /v1/orders/{order_id}/fulfillments:
    post:
      tags: [Fulfillments]
      operationId: createFulfillment
      deprecated: true
      summary: 历史 direct 商品履约接口
      description: 仅为旧订单兼容保留。当前四项 GPT 套餐均为 CDK 履约，新接入必须使用 /v1/redemptions 并选择 delivery_mode=auto_recharge 或 cdk。
      parameters:
        - $ref: '#/components/parameters/XPartnerId'
        - $ref: '#/components/parameters/XKeyId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XNonce'
        - $ref: '#/components/parameters/XSignature'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_data, customer_confirmed_email]
              properties:
                session_data:
                  type: object
                  description: 敏感字段；日志与错误响应不得回显
                  additionalProperties: true
                customer_confirmed_email: {type: boolean, const: true}
      responses:
        '202':
          description: 已受理
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Fulfillment'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Error'}
    get:
      tags: [Fulfillments]
      operationId: listFulfillments
      summary: 查询订单履约尝试
      parameters:
        - $ref: '#/components/parameters/XPartnerId'
        - $ref: '#/components/parameters/XKeyId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XNonce'
        - $ref: '#/components/parameters/XSignature'
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: 履约列表
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/Fulfillment'}}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    PartnerId: {type: apiKey, in: header, name: X-Partner-Id}
    KeyId: {type: apiKey, in: header, name: X-Key-Id}
    Timestamp: {type: apiKey, in: header, name: X-Timestamp}
    Nonce: {type: apiKey, in: header, name: X-Nonce}
    Signature: {type: apiKey, in: header, name: X-Signature}
  parameters:
    XPartnerId: {name: X-Partner-Id, in: header, required: true, schema: {type: string}}
    XKeyId: {name: X-Key-Id, in: header, required: true, schema: {type: string}}
    XTimestamp: {name: X-Timestamp, in: header, required: true, schema: {type: string, pattern: '^[0-9]{10}$'}}
    XNonce: {name: X-Nonce, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
    XSignature: {name: X-Signature, in: header, required: true, schema: {type: string, pattern: '^[a-f0-9]{64}$'}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 128}}
    OrderId:
      name: order_id
      in: path
      required: true
      schema: {type: string, pattern: '^ord_[A-Za-z0-9_-]+$'}
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    NotFound:
      description: 资源不存在或不属于当前代理商租户
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
...
    RedemptionEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [redemption_id, order_id, fulfillment_mode, status, failure_code, message, account_email_masked, fallback_recharge_available, created_at, finished_at]
          properties:
            redemption_id: {type: string}
            attempt_no: {type: integer, minimum: 1}
            progress_stage: {$ref: '#/components/schemas/ProgressStage'}
            progress_version: {type: integer, minimum: 0, description: 同一任务公开进度版本；历史记录未更新时为 0}
            progress_updated_at: {type: string, format: date-time, description: 最近进度变化时间；非每次查询时间}
            recovery_action: {type: [string, 'null'], enum: [retry, refund, null], description: 平台取消后的处理选择；退款关闭原单重提入口}
            error_category: {$ref: '#/components/schemas/RechargeErrorCategory'}
            retry_allowed: {type: boolean, description: 仅最新尝试明确失败且可安全重提时为 true}
            next_action: {type: string, enum: [wait, resubmit, none]}
            order_id: {type: string}
            fulfillment_mode: {type: string, enum: [direct, cdk]}
            status: {type: string, enum: [queued, running, succeeded, failed, cancelled]}
            failure_code: {type: [string, 'null']}
            message: {type: [string, 'null']}
            result_code: {type: [string, 'null'], description: 实际业务结果码的白标透传；进行中状态不得当作失败}
            result_stage: {type: [string, 'null'], description: 实际处理阶段；是否返回由平台按代理配置}
            card_last_four: {type: [string, 'null'], pattern: '^\d{4}$', description: 实际用卡后四位；绝不返回完整卡号}
            charge_amount_minor: {type: [integer, 'null'], minimum: 0, description: 实际扣款最小货币单位；是否返回由平台按代理配置}
            charge_currency: {type: [string, 'null']}
            upstream_order_id: {type: [string, 'null'], description: 内部履约参考号，默认隐藏；仅平台显式为该代理开放时返回}
            account_email_masked: {type: [string, 'null']}
            fallback_recharge_available: {type: boolean, description: 上游明确失败且允许重提后为 true；代理商在自己的订单页展示充值入口，不显示明文 CDK}
            created_at: {type: string, format: date-time}
            finished_at: {type: [string, 'null'], format: date-time}
...
    ProgressStage:
      type: string
      description: 默认公开的统一业务阶段；与敏感字段可见性无关，不包含上游身份或内部引用
      enum: [queued, logging_in, preparing_funds, awaiting_card, dispatching, processing, requires_action, confirming, review, upgrading, completed, failed, cancelled]
    RechargeErrorCategory:
      type: [string, 'null']
      enum: [credential, account, product, resource, service, confirmation, unknown, null]
      description: 脱敏业务分类，不是上游原始报文，也不代表重提授权；以 status/retry_allowed/next_action 为准
...
    Fulfillment:
      type: object
      required: [fulfillment_id, order_id, attempt_no, status, fulfillment_mode, created_at]
      properties:
        retry_allowed: {type: boolean}
        next_action: {type: string, enum: [wait, resubmit, none]}
        fulfillment_id: {type: string}
        order_id: {type: string}
        attempt_no: {type: integer, minimum: 1}
        progress_stage: {$ref: '#/components/schemas/ProgressStage'}
        progress_version: {type: integer, minimum: 0}
        progress_updated_at: {type: string, format: date-time}
        recovery_action: {type: [string, 'null'], enum: [retry, refund, null]}
        error_category: {$ref: '#/components/schemas/RechargeErrorCategory'}
        status: {type: string, enum: [queued, running, succeeded, failed, cancelled]}
        failure_code: {type: [string, 'null'], enum: [session_invalid, mailbox_login_failed, account_has_subscription, precheck_rejected, subscription_required, account_unavailable, product_unavailable, order_rejected, region_unsupported, payment_blocked, payment_declined, precharge_failed, cancelled, verification_timeout, service_unavailable, other, null]}
        message: {type: [string, 'null']}
        result_code: {type: [string, 'null'], description: 实际业务结果码，如 review/completed/declined}
        result_stage: {type: [string, 'null'], description: 实际处理阶段；是否返回由平台按代理配置}
        card_last_four: {type: [string, 'null'], pattern: '^\d{4}$', description: 实际用卡后四位；绝不返回完整卡号}
        charge_amount_minor: {type: [integer, 'null'], minimum: 0, description: 实际扣款最小货币单位}
        charge_currency: {type: [string, 'null']}
        upstream_order_id: {type: [string, 'null'], description: 内部履约参考号，默认隐藏；仅平台显式为该代理开放时返回}
        account_email_masked: {type: [string, 'null']}
        fulfillment_mode: {type: string, enum: [direct, cdk]}
        created_at: {type: string, format: date-time}
        finished_at: {type: [string, 'null'], format: date-time}
    RechargeCredential:
      oneOf:
        - type: object
          additionalProperties: false
          required: [mode, session]
          properties:
            mode: {type: string, const: session}
            session: {type: string, minLength: 1, maxLength: 32000, writeOnly: true}
        - type: object
          additionalProperties: false
          required: [mode, access_token]
          properties:
            mode: {type: string, const: access_token}
            access_token: {type: string, minLength: 1, maxLength: 32000, writeOnly: true}
        - type: object
          additionalProperties: false
          required: [mode, email, password]
          properties:
            mode: {type: string, const: mailbox}
            email: {type: string, format: email, maxLength: 320}
            password: {type: string, minLength: 1, maxLength: 1000, writeOnly: true}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}
            retryable: {type: boolean}